backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pytest==8.4.2
httpx==0.28.1
python-dotenv==1.0.1
langchain==0.3.27
langgraph==0.2.55
langchain-google-genai==3.0.3
google-genai>=1.0.0
pymongo[srv]>=4.6.0
